requirements.txt
fastapi==0.100.0
sqlalchemy==2.0.20
pydantic==2.0.2
python-dotenv==1.0.0
uvicorn==0.23.0
pyodbc==5.2.0
